[refactor] Move ``_compare_eq_mapping`` to its own module
