+++ src/_pytest/assertion/util.py
@@ -19,6 +19,7 @@ import _pytest._code
 from _pytest._io.pprint import PrettyPrinter
 from _pytest._io.saferepr import saferepr
 from _pytest._io.saferepr import saferepr_unlimited
+from _pytest.assertion._compare_mapping import _compare_eq_mapping
 from _pytest.assertion._compare_set import _compare_eq_set
 from _pytest.assertion._compare_set import SET_COMPARISON_FUNCTIONS
 from _pytest.assertion._typing import _HighlightFunc
@@ -490,52 +491,6 @@ def _compare_eq_sequence(
             explanation += [
                 f"{dir_with_more} contains {len_diff} more items, first extra item: {highlighter(extra)}"
             ]
-    return explanation
-
-
-def _compare_eq_mapping(
-    left: Mapping[object, object],
-    right: Mapping[object, object],
-    highlighter: _HighlightFunc,
-    verbose: int = 0,
-) -> list[str]:
-    explanation: list[str] = []
-    set_left = set(left)
-    set_right = set(right)
-    common = set_left.intersection(set_right)
-    same = {k: left[k] for k in common if left[k] == right[k]}
-    if same and verbose < 2:
-        explanation += [f"Omitting {len(same)} identical items, use -vv to show"]
-    elif same:
-        explanation += ["Common items:"]
-        explanation += highlighter(pprint.pformat(same)).splitlines()
-    diff = {k for k in common if left[k] != right[k]}
-    if diff:
-        explanation += ["Differing items:"]
-        for k in diff:
-            explanation += [
-                highlighter(saferepr({k: left[k]}))
-                + " != "
-                + highlighter(saferepr({k: right[k]}))
-            ]
-    extra_left = set_left - set_right
-    len_extra_left = len(extra_left)
-    if len_extra_left:
-        explanation.append(
-            f"Left contains {len_extra_left} more item{'' if len_extra_left == 1 else 's'}:"
-        )
-        explanation.extend(
-            highlighter(pprint.pformat({k: left[k] for k in extra_left})).splitlines()
-        )
-    extra_right = set_right - set_left
-    len_extra_right = len(extra_right)
-    if len_extra_right:
-        explanation.append(
-            f"Right contains {len_extra_right} more item{'' if len_extra_right == 1 else 's'}:"
-        )
-        explanation.extend(
-            highlighter(pprint.pformat({k: right[k] for k in extra_right})).splitlines()
-        )
     return explanation
